+++ src/main/java/io/beanmapper/exceptions/BeanConfigurationOperationNotAllowedException.java
@@ -1,6 +1,6 @@
 package io.beanmapper.exceptions;
 
-public class BeanConfigurationOperationNotAllowedException extends RuntimeException {
+public class BeanConfigurationOperationNotAllowedException extends BeanConfigurationException {
 
     public BeanConfigurationOperationNotAllowedException(String msg) {
         super(msg);